mailPart/requirements.txt
python-dotenv
EmailMessage